Multiple choice component feature (#9)
Any step with `choices` property is rendered as Multiple Choice Step. The `choices` is a list of `choice` where `choice` is a object containing `value` and `label`. `label` is what's displayed to the user and `value` is the value that represents a certain choice.

When user chooses choices at a Multiple Choice Step, the chosen list of values is stored at that step. Also, the step is replaced with a Text Step containing the text of space separated values.

A Multiple Choice Step can optionally contain `trigger` which, if `function` will receive list of `value`s of the chosen choices. If `trigger` is not a function, it is triggered without any further ado. One should note that `trigger` here is not the same as in the case of Options Step.

TODO: Make Multiple Choice Step variables updatable.